feat: activity report statistic:
