+++ packages/web/src/app/executive/activity-report/charged/[id]/page.tsx
@@ -8,7 +8,7 @@ import Custom404 from "@sparcs-clubs/web/app/not-found";
 import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import LoginRequired from "@sparcs-clubs/web/common/frames/LoginRequired";
 import { useAuth } from "@sparcs-clubs/web/common/providers/AuthContext";
-import ExecutiveActivityReportChargedFrame from "@sparcs-clubs/web/features/executive/activity-report/frames/ExecutiveActivityReportChargedFrame";
+import ExecutiveActivityReportChargedFrame from "@sparcs-clubs/web/features/activity-report/executive/frames/ExecutiveActivityReportChargedFrame";
 
 const ExecutiveActivityReport = () => {
   const { isLoggedIn, login, profile } = useAuth();
