+++ packages/web/src/app/executive/activity-report/club/[id]/page.tsx
@@ -11,8 +11,8 @@ import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import PageHead from "@sparcs-clubs/web/common/components/PageHead";
 import LoginRequired from "@sparcs-clubs/web/common/frames/LoginRequired";
 import { useAuth } from "@sparcs-clubs/web/common/providers/AuthContext";
+import ExecutiveActivityReportClubFrame from "@sparcs-clubs/web/features/activity-report/executive/frames/ExecutiveActivityReportClubFrame";
 import { useGetClubDetail } from "@sparcs-clubs/web/features/clubs/services/useGetClubDetail";
-import ExecutiveActivityReportClubFrame from "@sparcs-clubs/web/features/executive/activity-report/frames/ExecutiveActivityReportClubFrame";
 
 const ExecutiveActivityReportClub = () => {
   const { isLoggedIn, login, profile } = useAuth();
