+++ packages/web/src/features/activity-report/executive/frames/ExecutiveActivityReportClubFrame.tsx
@@ -1,12 +1,14 @@
 import { overlay } from "overlay-kit";
 import React, { useEffect, useState } from "react";
 
+import { ActivityStatusEnum } from "@clubs/interface/common/enum/activity.enum";
+
 import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import Button from "@sparcs-clubs/web/common/components/Button";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import SearchInput from "@sparcs-clubs/web/common/components/SearchInput";
 
-import ActivityReportClubStatistic from "../components/ActivityReportClubStatistic";
+import ActivityReportStatistic from "../components/ActivityReportStatistic";
 import ChargedChangeActivityModalContent from "../components/ChargedChangeActivityModalContent";
 import { ChargedChangeActivityProps } from "../components/ChargedChangeActivityModalTable";
 import ExecutiveClubActivitiesTable from "../components/ExecutiveClubActivitiesTable";
@@ -52,7 +54,24 @@ const ExecutiveActivityReportClubFrame: React.FC<{ clubId: string }> = ({
 
   return (
     <AsyncBoundary isLoading={isLoading} isError={isError}>
-      <ActivityReportClubStatistic data={data ?? { items: [] }} />
+      <ActivityReportStatistic
+        pendingTotalCount={
+          data?.items.filter(
+            item => item.activityStatusEnum === ActivityStatusEnum.Applied,
+          ).length ?? 0
+        }
+        approvedTotalCount={
+          data?.items.filter(
+            item => item.activityStatusEnum === ActivityStatusEnum.Approved,
+          ).length ?? 0
+        }
+        rejectedTotalCount={
+          data?.items.filter(
+            item => item.activityStatusEnum === ActivityStatusEnum.Rejected,
+          ).length ?? 0
+        }
+        withApprovedRate
+      />
       <FlexWrapper direction="row" gap={16}>
         <SearchInput
           searchText={searchText}
